pyproject.toml
[project]
name = "e-commerce-fastapi"
version = "0.1.0"
description = "E-commerce backend API built with FastAPI, SQLAlchemy, bearer authentication, and Alembic migrations."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.1",
    "fastapi>=0.121.1",
    "passlib>=1.7.4",
    "pip>=25.3",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
